+++ lib/apathy_drive/systems/exits/command.ex
@@ -5,7 +5,7 @@ defmodule Systems.Exits.Command do
     room_exit["name"]
   end
 
-  def move(spirit, nil, current_room, room_exit)  do
+  def move(spirit, nil, _current_room, _room_exit)  do
     send_message(spirit, "scroll", "<p>There is no exit in that direction.</p>")
   end
 
@@ -13,7 +13,7 @@ defmodule Systems.Exits.Command do
     move_via_command(nil, monster, current_room, room_exit)
   end
 
-  def move(spirit, monster, current_room, room_exit) do
+  def move(spirit, _monster, _current_room, _room_exit) do
     send_message(spirit, "scroll", "<p>There is no exit in that direction.</p>")
   end
 
@@ -28,6 +28,39 @@ defmodule Systems.Exits.Command do
 
     Components.Hints.deactivate(spirit, "movement")
     Systems.Room.display_room_in_scroll(spirit, destination)
+  end
+
+  def move_via_command(nil, monster, current_room, room_exit) do
+    if !Systems.Combat.stunned?(monster) do
+      destination = Rooms.find_by_id(room_exit["destination"])
+      Components.Monsters.remove_monster(current_room, monster)
+      Components.Monsters.add_monster(destination, monster)
+      Entities.save!(destination)
+      Entities.save!(current_room)
+      Entities.save(monster)
+
+      if room_exit["from_message"] do
+        Systems.Monster.observers(current_room, monster)
+        |> Enum.each(fn(observer) ->
+          send_message(observer, "scroll", "<p><span class='dark-green'>#{interpolate(room_exit["from_message"], %{"user" => monster})}</span></p>")
+        end)
+      else
+        Systems.Monster.display_exit_message(current_room, monster)
+      end
+
+      if room_exit["to_message"] do
+        Systems.Monster.observers(destination, monster)
+        |> Enum.each(fn(observer) ->
+          send_message(observer, "scroll", "<p><span class='dark-green'>#{interpolate(room_exit["to_message"], %{"user" => monster})}</span></p>")
+        end)
+      else
+        Systems.Monster.display_enter_message(destination, monster)
+      end
+
+      Systems.Aggression.monster_entered(monster, destination)
+
+      Systems.Room.display_room_in_scroll(monster, destination)
+    end
   end
 
   def move_via_command(spirit, monster, current_room, room_exit) do
@@ -49,7 +82,7 @@ defmodule Systems.Exits.Command do
       if room_exit["from_message"] do
         Systems.Monster.observers(current_room, monster)
         |> Enum.each(fn(observer) ->
-          send_message(observer, "scroll", "<p><span class='yellow'>#{interpolate(room_exit["from_message"], %{"user" => monster})}</span></p>")
+          send_message(observer, "scroll", "<p><span class='dark-green'>#{interpolate(room_exit["from_message"], %{"user" => monster})}</span></p>")
         end)
       else
         Systems.Monster.display_exit_message(current_room, monster)
@@ -58,7 +91,7 @@ defmodule Systems.Exits.Command do
       if room_exit["to_message"] do
         Systems.Monster.observers(destination, monster)
         |> Enum.each(fn(observer) ->
-          send_message(observer, "scroll", "<p><span class='yellow'>#{interpolate(room_exit["to_message"], %{"user" => monster})}</span></p>")
+          send_message(observer, "scroll", "<p><span class='dark-green'>#{interpolate(room_exit["to_message"], %{"user" => monster})}</span></p>")
         end)
       else
         Systems.Monster.display_enter_message(destination, monster)
@@ -71,7 +104,7 @@ defmodule Systems.Exits.Command do
     end
   end
 
-  def look(spirit, monster, current_room, room_exit) do
+  def look(spirit, _monster, _current_room, _room_exit) do
     send_message(spirit, "scroll", "<p>There is no exit in that direction.</p>")
   end