+++ lib/apathy_drive/systems/exits/trap.ex
@@ -71,12 +71,12 @@ defmodule Systems.Exits.Trap do
 
     Systems.Monster.observers(current_room, monster)
     |> Enum.each(fn(observer) ->
-      send_message(observer, "scroll", "<p><span class='yellow'>#{interpolate(room_exit["from_message"], %{"user" => monster})}</span></p>")
+      send_message(observer, "scroll", "<p><span class='dark-green'>#{interpolate(room_exit["from_message"], %{"user" => monster})}</span></p>")
     end)
 
     Systems.Monster.observers(destination, monster)
     |> Enum.each(fn(observer) ->
-      send_message(observer, "scroll", "<p><span class='yellow'>#{interpolate(room_exit["to_message"], %{"user" => monster})}</span></p>")
+      send_message(observer, "scroll", "<p><span class='dark-green'>#{interpolate(room_exit["to_message"], %{"user" => monster})}</span></p>")
     end)
 
     amount = monster