+++ lib/apathy_drive/systems/monster.ex
@@ -2,6 +2,7 @@ defmodule Systems.Monster do
   use Systems.Reload
   import Systems.Text
   import Utility
+  import BlockTimer
   use Timex
 
   def spawn_monster(monster) do
@@ -40,6 +41,7 @@ defmodule Systems.Monster do
       equip_monster(entity)
       Components.Abilities.reset_abilities(entity)
       Entity.add_to_type_collection(entity)
+
       entity
     end
   end
@@ -50,6 +52,15 @@ defmodule Systems.Monster do
       Components.Monsters.add_monster(room, monster)
 
       display_enter_message(room, monster)
+
+      if Components.Characters.get_characters(room) |> Enum.any? do
+        Systems.Monster.monsters_in_room(room, monster)
+        |> Enum.each(fn(monster_in_room) ->
+             Systems.Aggression.monster_entered(monster_in_room, room)
+           end)
+
+        Systems.Aggression.monster_entered(monster, room)
+      end
     end
   end
 
@@ -79,7 +90,7 @@ defmodule Systems.Monster do
 
     observers(room, monster)
     |> Enum.each(fn(observer) ->
-      send_message(observer,"scroll", "<p><span class='yellow'>#{message}</span></p>")
+      send_message(observer,"scroll", "<p><span class='dark-green'>#{message}</span></p>")
     end)
   end
 
@@ -93,7 +104,7 @@ defmodule Systems.Monster do
 
     observers(room, monster)
     |> Enum.each(fn(observer) ->
-      send_message(observer,"scroll", "<p><span class='yellow'>#{message}</span></p>")
+      send_message(observer,"scroll", "<p><span class='dark-green'>#{message}</span></p>")
     end)
   end
 
@@ -107,7 +118,7 @@ defmodule Systems.Monster do
 
     observers(room, monster)
     |> Enum.each(fn(observer) ->
-      send_message(observer, "scroll", "<p><span class='yellow'>#{message}</span></p>")
+      send_message(observer, "scroll", "<p><span class='dark-green'>#{message}</span></p>")
     end)
   end
 
@@ -121,7 +132,7 @@ defmodule Systems.Monster do
 
     observers(room, monster)
     |> Enum.each(fn(observer) ->
-      send_message(observer, "scroll", "<p><span class='yellow'>#{message}</span></p>")
+      send_message(observer, "scroll", "<p><span class='dark-green'>#{message}</span></p>")
     end)
   end
 
@@ -217,8 +228,8 @@ defmodule Systems.Monster do
 
       def description,   do: nil
       def death_message, do: ~s({{name}} drops dead before you.)
-      def enter_message, do: ~s({{name}} walks in from {{direction}}.)
-      def exit_message,  do: ~s(A {{name}} walks off to %s.)
+      def enter_message, do: ~s(<span class='yellow'>{{name}}</span> walks in from {{direction}}.)
+      def exit_message,  do: ~s(<span class='yellow'>{{name}}</span> walks off to {{direction}}.)
       def abilities,     do: []
       def greeting,      do: "The #{name} completely ignores you."
       def gender,        do: nil